Rework RecentOrders screen to improve look and feel
Changed the buttons for 'Add new order' and 'search' to use just icons
instead of buttons with text.
Each order in the ListView is clickable, and clicking an order will expand
it to show more of the details, like the number of each type of item purchased.
Also included an edit button in the expanded view so the user can edit the
order if needed.