Allow user to select items on OrderDetails screen
Added code to allow user to select individual line items on the order
details screen.  When they select an item, an icon (solid circle) will
appear to the left of the item.  For the other non-selected items,
an empty circle will appear to the left of the item.

When the user selects a non-selected item, the icon will change from
an empty circle to a solid circle.  Once no items are selected at all,
the circle icons will disappear for all the rows. If the user clicks
on an item again, then the circle icons will re-appear.